+++ perl/FSM/HDL/FlattenedDT.pm
@@ -24,6 +24,7 @@ use FSM::Synthesis::EnableGraph::SignalSupport;
 use FSM::HDL::FlattenedDT::Orchestrator;
 use FSM::HDL::FlattenedDT::Backend::SystemVerilog::ASTFactorizationSupport;
 use FSM::HDL::FlattenedDT::Backend::SystemVerilog::ConsolidatedIntermediateEmitter;
+use FSM::HDL::FlattenedDT::Backend::SystemVerilog::GlobalFactorizationSupport;
 use FSM::HDL::FlattenedDT::Backend::SystemVerilog::InternalDeclarationEmitter;
 use FSM::HDL::FlattenedDT::Backend::SystemVerilog::IntermediateSignalSupport;
 use FSM::HDL::FlattenedDT::Backend::SystemVerilog::ScaffoldEmitter;
@@ -103,6 +104,7 @@ sub new ($class, %args) {
     $self->{backend_sv_scaffold} = FSM::HDL::FlattenedDT::Backend::SystemVerilog::ScaffoldEmitter->new(flattened_dt => $self);
     $self->{backend_sv_internal_decl} = FSM::HDL::FlattenedDT::Backend::SystemVerilog::InternalDeclarationEmitter->new(flattened_dt => $self);
     $self->{backend_sv_intermediate_support} = FSM::HDL::FlattenedDT::Backend::SystemVerilog::IntermediateSignalSupport->new(flattened_dt => $self);
+    $self->{backend_sv_global_factorization} = FSM::HDL::FlattenedDT::Backend::SystemVerilog::GlobalFactorizationSupport->new(flattened_dt => $self);
     $self->{backend_sv_ast_factorization} = FSM::HDL::FlattenedDT::Backend::SystemVerilog::ASTFactorizationSupport->new(flattened_dt => $self);
     $self->{backend_sv_consolidated_intermediate} = FSM::HDL::FlattenedDT::Backend::SystemVerilog::ConsolidatedIntermediateEmitter->new(flattened_dt => $self);
     $self->{backend_verilog} = FSM::HDL::FlattenedDT::Backend::Verilog->new(flattened_dt => $self);